[TASK] move ConfigLoader into Application
The configuration loader class was placed inside the commands namespace.
As the configuration is more application related than it is a command
(SOLID) it moves into the application namespace.

Next to that this also prepares moving configuration related task from the
Application god-like-class.
