+++ tests/N98/Magento/ApplicationTest.php
@@ -162,7 +162,7 @@ class ApplicationTest extends TestCase
         );
 
         $configurationLoader = $this->getMock(
-            '\N98\Magento\Command\ConfigurationLoader',
+            '\N98\Magento\Application\ConfigurationLoader',
             array('getConfigurationLoaderDir'),
             array(array(), false, new NullOutput())
         );
